+++ packages/megadoc-html-serializer/lib/lintingRules.js
@@ -2,7 +2,7 @@ const { LOG_ERROR } = require('megadoc-linter')
 
 exports.NoBrokenLinks = {
   name: 'html/no-broken-links',
-  messageFn: ({ path }) => `Unable to resolve link to "${path}"`,
+  messageFn: ({ path }) => `"${path}" not found`,
   defaults: {
     level: LOG_ERROR
   }