+++ packages/megadoc-plugin-js/lib/__tests__/reduceTreeFn.test.js
@@ -4,7 +4,7 @@ const b = require('megadoc-corpus').builders;
 
 describe('megadoc-plugin-js::reduceTreeFn', function() {
   const defaultContext = {
-    commonOptions: {},
+    compilerOptions: {},
     options: {},
   };